Compilation & Dependency Issues Fix with Swagger running #1 (#29)

+++ src/main/java/com/iemr/ecd/repo/call_conf_allocation/QuestionnaireSectionRepo.java
@@ -29,7 +29,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.ecd.dao.QuestionnaireSections;
 
 @Repository
-public interface QuestionnaireSectionRepo extends CrudRepository<QuestionnaireSections, Long> {
+public interface QuestionnaireSectionRepo extends CrudRepository<QuestionnaireSections, Integer> {
 
 	List<QuestionnaireSections> findByPsmIdOrderByLastModDateDesc(Integer psmIds);
